backend/pyproject.toml
[project]
name = "lawyer-backend"
version = "0.1.0"
description = "Backend API for Lawyer Office SaaS"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "python-multipart>=0.0.9",
]

[tool.uvicorn]
factory = false

